Feat: abstracting queries into it's own module for  testing production and development
